// src/api/service/model/noteDto.ts
/**
 * Things Keeper
 * Will save your notes.
 *
 * OpenAPI spec version: 0.1.0
 * Contact: vrahlinskiy@cinimex.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Tag } from './tag';


export interface NoteDto { 
    created?: Date;
    description?: string;
    id?: string;
    tags?: Array<Tag>;
    title?: string;
}